// src/Models/cardModel.ts
export interface cardModel {
    id?: number;
    userid?: string;
    businessName?: string;
    businessDescription?: string;
    image?: any;
    phone?: string
    email?: string;
    instagram?: string;
    facebook?: string;
    twitter?: string;
    github?: string;
    location?: string;
    website?: string;
    templateNum?: number;
    upload? : any;
    iconsColor?: string;
}